requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Data Processing
numpy<2.0,>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Model Evaluation
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Explainability/Interpretability
grad-cam>=1.5.2
captum>=0.6.0
shap>=0.42.0
lime>=0.2.0.1
eli5>=0.13.0

# Graph Neural Networks
torch-geometric>=2.3.0
networkx>=3.1

# Medical Imaging
pydicom>=2.4.0
nibabel>=5.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
kaggle>=1.5.16

# API & Deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.3
streamlit-aggrid>=0.3.4

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.7.0
flake8>=6.1.0
pytest>=7.4.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.36.0
